Fix cached decorations (#802)
* computeViewRange should not be calling setupFromView.

* DecorationsCache - prohibits removal during decorate.

* Rename

* Tests.

Clean up other tests.

=Deprecate DecorateContext.screenViewport.

* lint, api, docs.

* Update .md links.
