+++ common/api/summary/imodeljs-frontend.exports.csv
@@ -107,6 +107,7 @@ alpha;CustomQuantityTypeDefinition
 internal;DebugShaderFile
 public;DecorateContext 
 public;Decorations 
+internal;DecorationsCache
 public;Decorator 
 internal;DefaultViewTouchTool 
 internal;DefineACSByElementTool
