+++ common/changes/@bentley/imodeljs-frontend/fix-decoration-cache_2021-02-19-11-11.json
@@ -1,0 +1,11 @@
+{
+  "changes": [
+    {
+      "packageName": "@bentley/imodeljs-frontend",
+      "comment": "Prevent buggy decorators from invalidating other decorators' cached decorations.",
+      "type": "none"
+    }
+  ],
+  "packageName": "@bentley/imodeljs-frontend",
+  "email": "22944042+pmconne@users.noreply.github.com"
+}
