+++ core/frontend/src/Viewport.ts
@@ -36,12 +36,11 @@ import { IModelConnection } from "./IModelConnection";
 import { linePlaneIntersect } from "./LinePlaneIntersect";
 import { ToolTipOptions } from "./NotificationManager";
 import { PerModelCategoryVisibility } from "./PerModelCategoryVisibility";
-import { CanvasDecoration } from "./render/CanvasDecoration";
 import { Decorations } from "./render/Decorations";
 import { FeatureSymbology } from "./render/FeatureSymbology";
 import { GraphicType } from "./render/GraphicBuilder";
 import { Pixel } from "./render/Pixel";
-import { GraphicList, RenderGraphicOwner } from "./render/RenderGraphic";
+import { GraphicList } from "./render/RenderGraphic";
 import { RenderMemory } from "./render/RenderMemory";
 import { createRenderPlanFromViewport } from "./render/RenderPlan";
 import { RenderTarget } from "./render/RenderTarget";
@@ -59,6 +58,7 @@ import { ViewPose } from "./ViewPose";
 import { ViewRect } from "./ViewRect";
 import { ModelDisplayTransformProvider, ViewState } from "./ViewState";
 import { ViewStatus } from "./ViewStatus";
+import { DecorationsCache } from "./DecorationsCache";
 
 // cSpell:Ignore rect's ovrs subcat subcats unmounting UI's
 
@@ -82,17 +82,6 @@ export interface ViewportDecorator {
    * @see [[useCachedDecorations]] to avoid unnecessarily recreating decorations.
    */
   decorate(context: DecorateContext): void;
-}
-
-/** @internal */
-export type CachedDecoration =
-  { type: "graphic", graphicType: GraphicType, graphicOwner: RenderGraphicOwner } |
-  { type: "canvas", canvasDecoration: CanvasDecoration, atFront: boolean } |
-  { type: "html", htmlElement: HTMLElement };
-
-function disposeCachedDecoration(dec: CachedDecoration): void {
-  if ("graphic" === dec.type)
-    dec.graphicOwner.disposeGraphic();
 }
 
 /** @alpha Source of depth point returned by [[Viewport.pickDepthPoint]]. */
@@ -291,7 +280,6 @@ export abstract class Viewport implements IDisposable {
   private _timePointValid = false;
   /** @internal */
   public get timePointValid() { return this._timePointValid; }
-  private readonly _decorationCache: Map<ViewportDecorator, CachedDecoration[]> = new Map<ViewportDecorator, CachedDecoration[]>();
 
   /** Strictly for tests. @internal */
   public setAllValid(): void {
@@ -314,13 +302,14 @@ export abstract class Viewport implements IDisposable {
     this._sceneValid = false;
     this._timePointValid = false;
     this.invalidateDecorations();
-    this._disposeDecorationCache(); // When the scene is invalidated, remove all cached decorations so they get regenerated.
   }
+
   /** @internal */
   public invalidateRenderPlan(): void {
     this._renderPlanValid = false;
     this.invalidateScene();
   }
+
   /** @internal */
   public invalidateController(): void {
     this._controllerValid = this._analysisFractionValid = false;
@@ -331,6 +320,7 @@ export abstract class Viewport implements IDisposable {
   public setValidScene() {
     this._sceneValid = true;
   }
+
   /** @deprecated Use requestRedraw.
    * @internal
    */
@@ -1023,40 +1013,10 @@ export abstract class Viewport implements IDisposable {
     IModelApp.tileAdmin.registerViewport(this);
   }
 
-  /** Forces removal of a specific decorator's cached decorations from this viewport, if they exist.
-   * This will force those decorations to be regenerated.
-   * @see [[ViewportDecorator.useCachedDecorations]].
-   * @beta
-   */
-  public invalidateCachedDecorations(decorator: ViewportDecorator) {
-    assert(true === decorator.useCachedDecorations, "Cannot invalidate cached decorations on a decorator which does not support cached decorations.");
-    if (true !== decorator.useCachedDecorations)
-      return;
-
-    const decorations = this._decorationCache.get(decorator);
-    if (decorations) {
-      decorations.forEach((decoration) => disposeCachedDecoration(decoration));
-      this._decorationCache.delete(decorator);
-    }
-
-    // Always invalidate decorations. Decorator may have no cached decorations currently, but wants them created.
-    this.invalidateDecorations();
-  }
-
-  /** Disposes of the entire decoration cache for this viewport. */
-  private _disposeDecorationCache() {
-    this._decorationCache.forEach((decorations: CachedDecoration[]) => {
-      decorations.forEach((decoration) => disposeCachedDecoration(decoration));
-    });
-
-    this._decorationCache.clear();
-  }
-
   public dispose(): void {
     if (this.isDisposed)
       return;
 
-    this._disposeDecorationCache();
     this._target = dispose(this._target);
     this.subcategories.dispose();
     IModelApp.tileAdmin.forgetViewport(this);
@@ -1944,7 +1904,6 @@ export abstract class Viewport implements IDisposable {
 
   /** @internal */
   public computeViewRange(): Range3d {
-    this.setupFromView(); // can't proceed if viewport isn't valid (not active)
     const fitRange = this.view.computeFitRange();
     this.forEachTiledGraphicsProviderTree((ref) => {
       ref.unionFitRange(fitRange);
@@ -2312,23 +2271,6 @@ export abstract class Viewport implements IDisposable {
   /** @internal */
   protected addDecorations(_decorations: Decorations): void { }
 
-  /** @internal */
-  public getCachedDecorations(decorator: ViewportDecorator): CachedDecoration[] | undefined {
-    return this._decorationCache.get(decorator);
-  }
-
-  /** @internal */
-  public addCachedDecoration(decorator: ViewportDecorator, decoration: CachedDecoration): void {
-    assert(true === decorator.useCachedDecorations);
-    let list = this._decorationCache.get(decorator);
-    if (!list) {
-      list = [];
-      this._decorationCache.set(decorator, list);
-    }
-
-    list.push(decoration);
-  }
-
   /** Read selected data about each pixel within a rectangular region of this Viewport.
    * @param rect The area of the viewport's contents to read. The origin specifies the upper-left corner. Must lie entirely within the viewport's dimensions. This input viewport is specified using CSS pixels not device pixels.
    * @param selector Specifies which aspect(s) of data to read.
@@ -2590,6 +2532,7 @@ export class ScreenViewport extends Viewport {
   private _lastPose?: ViewPose; // the pose the last time this view was rendered
   private _webglCanvas?: HTMLCanvasElement;
   private _logo!: HTMLImageElement;
+  private readonly _decorationCache = new DecorationsCache();
 
   /** The parent HTMLDivElement of the canvas. */
   public readonly parentDiv: HTMLDivElement;
@@ -2621,6 +2564,32 @@ export class ScreenViewport extends Viewport {
     const vp = new this(canvas, parentDiv, IModelApp.renderSystem.createTarget(canvas));
     vp.changeView(view);
     return vp;
+  }
+
+  /** @internal */
+  public dispose(): void {
+    super.dispose();
+    this._decorationCache.clear();
+  }
+
+  /** @internal */
+  public invalidateScene(): void {
+    super.invalidateScene();
+
+    // When the scene is invalidated, so are all cached decorations - they will be regenerated.
+    this._decorationCache.clear();
+  }
+
+  /** Forces removal of a specific decorator's cached decorations from this viewport, if they exist.
+   * This will force those decorations to be regenerated.
+   * @see [[ViewportDecorator.useCachedDecorations]].
+   * @beta
+   */
+  public invalidateCachedDecorations(decorator: ViewportDecorator) {
+    this._decorationCache.delete(decorator);
+
+    // Always invalidate decorations. Decorator may have no cached decorations currently, but wants them created.
+    this.invalidateDecorations();
   }
 
   /** @internal */
@@ -2900,15 +2869,23 @@ export class ScreenViewport extends Viewport {
     // SEE: decorationDiv doc comment
     // eslint-disable-next-line deprecation/deprecation
     ScreenViewport.markAllChildrenForRemoval(this.decorationDiv);
-    const context = new DecorateContext(this, decorations);
-    context.addFromDecorator(this.view);
-    this.forEachTiledGraphicsProviderTree((ref) => context.addFromDecorator(ref));
+    const context = new DecorateContext(this, decorations, this._decorationCache);
+    try {
+      // It is an error to try to remove cached decorations while we are decorating.
+      // Some naughty decorators unwittingly do so by e.g. invalidating the scene in their decorate method.
+      this._decorationCache.prohibitRemoval = true;
 
-    for (const decorator of IModelApp.viewManager.decorators)
-      context.addFromDecorator(decorator);
+      context.addFromDecorator(this.view);
+      this.forEachTiledGraphicsProviderTree((ref) => context.addFromDecorator(ref));
 
-    // eslint-disable-next-line deprecation/deprecation
-    ScreenViewport.removeMarkedChildren(this.decorationDiv);
+      for (const decorator of IModelApp.viewManager.decorators)
+        context.addFromDecorator(decorator);
+
+      // eslint-disable-next-line deprecation/deprecation
+      ScreenViewport.removeMarkedChildren(this.decorationDiv);
+    } finally {
+      this._decorationCache.prohibitRemoval = false;
+    }
   }
 
   /** Change the cursor for this Viewport */
