+++ docs/changehistory/2.5.0.md
@@ -87,7 +87,7 @@ class FancyDecorator {
 }
 ```
 
-[ViewManager.invalidateCachedDecorationsAllViews]($frontend) (and [Viewport.invalidateCachedDecorations]($frontend)) give the decorator much tighter control over when its decorations are regenerated, potentially resulting in significantly improved performance.
+[ViewManager.invalidateCachedDecorationsAllViews]($frontend) (and [ScreenViewport.invalidateCachedDecorations]($frontend)) give the decorator much tighter control over when its decorations are regenerated, potentially resulting in significantly improved performance.
 
 ## FeatureSymbology namespace
